Move check_username_availability rate limiter to rate_limiter.py
This pattern is used in other apps with rate limiters
